Move all code to punctuate_capitalize.py
