+++ nemo/collections/nlp/models/token_classification/punctuation_capitalization_model.py
@@ -15,7 +15,6 @@
 import os
 from typing import Dict, List, Optional
 
-import numpy as np
 import torch
 from omegaconf import DictConfig, OmegaConf
 from pytorch_lightning import Trainer
@@ -29,8 +28,9 @@ from nemo.collections.nlp.metrics.classification_report import ClassificationRep
 from nemo.collections.nlp.models.nlp_model import NLPModel
 from nemo.collections.nlp.modules.common import TokenClassifier
 from nemo.collections.nlp.modules.common.lm_utils import get_lm_model
+from nemo.collections.nlp.parts.utils_funcs import tensor2list
 from nemo.core.classes.common import PretrainedModelInfo, typecheck
-from nemo.core.classes.exportable import Exportable
+from nemo.core.classes.exportable import Exportable, ExportFormat
 from nemo.core.neural_types import LogitsType, NeuralType
 from nemo.utils import logging
 
@@ -335,7 +335,7 @@ class PunctuationCapitalizationModel(NLPModel, Exportable):
         )
 
     def _setup_infer_dataloader(
-        self, queries: List[str], batch_size: int, max_seq_length: int = None, step: int = None, margin: int = None
+        self, queries: List[str], batch_size: int, max_seq_length: int = None
     ) -> 'torch.utils.data.DataLoader':
         """
         Setup function for a infer data loader.
@@ -349,13 +349,9 @@ class PunctuationCapitalizationModel(NLPModel, Exportable):
         """
         if max_seq_length is None:
             max_seq_length = self._cfg.dataset.max_seq_length
-        if step is None:
-            step = self._cfg.dataset.step
-        if margin is None:
-            margin = self._cfg.dataset.margin
 
         dataset = BertPunctuationCapitalizationInferDataset(
-            tokenizer=self.tokenizer, queries=queries, max_seq_length=max_seq_length, step=step, margin=margin
+            tokenizer=self.tokenizer, queries=queries, max_seq_length=max_seq_length
         )
 
         return torch.utils.data.DataLoader(
@@ -368,163 +364,92 @@ class PunctuationCapitalizationModel(NLPModel, Exportable):
             drop_last=False,
         )
 
-    @staticmethod
-    def move_acc_probs_to_token_preds(pred, acc_prob, number_of_probs_to_move):
-        if number_of_probs_to_move > acc_prob.shape[0]:
-            raise ValueError(
-                f"Not enough accumulated probabilities. Number_of_probs_to_move={number_of_probs_to_move} "
-                f"acc_prob.shape={acc_prob.shape}"
-            )
-        if number_of_probs_to_move > 0:
-            pred = pred + list(np.argmax(acc_prob[:number_of_probs_to_move], axis=-1))
-        acc_prob = acc_prob[number_of_probs_to_move:]
-        return pred, acc_prob
-
-    @staticmethod
-    def update_accumulated_probabilities(acc_prob, update):
-        acc_prob *= update[: acc_prob.shape[0]]
-        acc_prob = np.concatenate([acc_prob, update[acc_prob.shape[0] :]], axis=0)
-        return acc_prob
-
-    @staticmethod
-    def remove_margins(tensor, margin_size, keep_left, keep_right):
-        if not keep_left:
-            tensor = tensor[margin_size + 1 :]  # remove left margin and CLS token
-        if not keep_right:
-            tensor = tensor[: tensor.shape[0] - margin_size - 1]  # remove right margin and SEP token
-        return tensor
-
-    def apply_punct_capit_predictions(self, query, punct_preds, capit_preds):
-        query = query.strip().split()
-        assert len(query) == len(
-            punct_preds
-        ), f"len(query)={len(query)} len(punct_preds)={len(punct_preds)}, query[:30]={query[:30]}"
-        assert len(query) == len(
-            capit_preds
-        ), f"len(query)={len(query)} len(capit_preds)={len(capit_preds)}, query[:30]={query[:30]}"
-        punct_ids_to_labels = {v: k for k, v in self._cfg.punct_label_ids.items()}
-        capit_ids_to_labels = {v: k for k, v in self._cfg.capit_label_ids.items()}
-        query_with_punct_and_capit = ''
-        for j, word in enumerate(query):
-            punct_label = punct_ids_to_labels[punct_preds[j]]
-            capit_label = capit_ids_to_labels[capit_preds[j]]
-
-            if capit_label != self._cfg.dataset.pad_label:
-                word = word.capitalize()
-            query_with_punct_and_capit += word
-            if punct_label != self._cfg.dataset.pad_label:
-                query_with_punct_and_capit += punct_label
-            query_with_punct_and_capit += ' '
-        return query_with_punct_and_capit[:-1]
-
-    def _transform_logit_to_prob_and_remove_margins_and_extract_word_probs(
-        self, punct_logits, capit_logits, subtokens_mask, start_word_ids, margin, is_first, is_last, query_ids,
-    ):
-        new_start_word_ids = list(start_word_ids)
-        subtokens_mask = subtokens_mask > 0.5
-        b_punct_probs, b_capit_probs = [], []
-        for i, (first, last, q_i, pl, cl, stm) in enumerate(
-            zip(is_first, is_last, query_ids, punct_logits, capit_logits, subtokens_mask)
-        ):
-            if not first:
-                new_start_word_ids[i] += torch.count_nonzero(stm[: margin + 1]).numpy()  # + 1 is for [CLS] token
-            stm = self.remove_margins(stm, margin, keep_left=first, keep_right=last)
-            for b_probs, logits in [(b_punct_probs, pl), (b_capit_probs, cl)]:
-                p = torch.nn.functional.softmax(
-                    self.remove_margins(logits, margin, keep_left=first, keep_right=last)[stm], dim=-1,
-                )
-                b_probs.append(p.detach().cpu().numpy())
-        return b_punct_probs, b_capit_probs, new_start_word_ids
-
     def add_punctuation_capitalization(
-        self, queries: List[str], batch_size: int = None, max_seq_length: int = 64, step: int = 8, margin: int = 16
+        self, queries: List[str], batch_size: int = None, max_seq_length: int = 512
     ) -> List[str]:
         """
-        Adds punctuation and capitalization to the queries. Use this method for inference.
-
-        Parameters ``max_seq_length``, ``step``, ``margin`` are for controlling the way queries are split into segments
-        which then processed by the model. Parameter ``max_seq_length`` is a length of a segment after tokenization
-        including special tokens [CLS] in the beginning and [SEP] in the end of a segment. Parameter ``step`` is shift
-        between consequent segments. Parameter ``margin`` is used to exclude negative effect of subtokens near
-        borders of segments which have only one side context.
-
-        If segments overlap, probabilities of overlapping predictions are multiplied and then the label with
-        corresponding to the maximum probability is selected.
-
+        Adds punctuation and capitalization to the queries. Use this method for debugging and prototyping.
         Args:
             queries: lower cased text without punctuation
             batch_size: batch size to use during inference
-            max_seq_length: maximum sequence length of segment after tokenization.
-            step: relative shift of consequent segments into which long queries are split. Long queries are split into
-                segments which can overlap. Parameter ``step`` controls such overlapping. Imagine that queries are
-                tokenized into characters, ``max_seq_length=5``, and ``step=2``. In such a case query "hello" is
-                tokenized into segments ``[['[CLS]', 'h', 'e', 'l', '[SEP]'], ['[CLS]', 'l', 'l', 'o', '[SEP]']]``.
-            margin: number of subtokens in the beginning and the end of segments which are not used for prediction
-                computation. The first segment does not have left margin and the last segment does not have right
-                margin. For example, if input sequence is tokenized into characters, ``max_seq_length=5``,
-                ``step=1``, and ``margin=1`` than query "hello" will be tokenized into segments
-                ``[['[CLS]', 'h', 'e', 'l', '[SEP]'], ['[CLS]', 'e', 'l', 'l', '[SEP]'],
-                ['[CLS]', 'l', 'l', 'o', '[SEP]']]``. These segments are passed to the model. Before final predictions
-                computation, margins are removed. In the next list, subtokens which logits are not used for final
-                predictions computation are marked with asterisk: ``[['[CLS]'*, 'h', 'e', 'l'*, '[SEP]'*],
-                ['[CLS]'*, 'e'*, 'l', 'l'*, '[SEP]'*], ['[CLS]'*, 'l'*, 'l', 'o', '[SEP]'*]]``.
+            max_seq_length: maximum sequence length after tokenization
         Returns:
-            result: text with added capitalization and punctuation ``max_seq_length`` equals 5, ``step`` equals 2, and
-
+            result: text with added capitalization and punctuation
         """
-        if len(queries) == 0:
+        if queries is None or len(queries) == 0:
             return []
         if batch_size is None:
             batch_size = len(queries)
             logging.info(f'Using batch size {batch_size} for inference')
-        result = []
-        mode = self.training
-        d = 'cuda' if torch.cuda.is_available() else 'cpu'
-        try:
-            self.eval()
-            self = self.to(d)
-            infer_datalayer = self._setup_infer_dataloader(queries, batch_size, max_seq_length, step, margin)
 
-            # Predicted labels for queries. List of labels for every query
-            all_punct_preds, all_capit_preds = [[] for _ in queries], [[] for _ in queries]
-            # Accumulated probabilities (or product of probabilities acquired from different segments) of punctuation
-            # and capitalization. Probabilities for words in query are extracted using `subtokens_mask`.Probabilities
-            # for newly processed words are appended to the accumulated probabilities. If probabilities for a word are
-            # already present in `acc_probs`, old probabilities are replaced with multiplication of old probabilities
-            # and probabilities acquired from new segment. Segments are processed in the order they are present in an
-            # input query. When all segments with a word are processed, the label with highest probability is chosen
-            # and appended to an appropriate list in `all_preds`. After adding prediction to `all_preds`,
-            # probabilities for a word are removed from `acc_probs`
-            acc_punct_probs, acc_capit_probs = [None for _ in queries], [None for _ in queries]
-            for batch_i, batch in enumerate(infer_datalayer):
-                inp_ids, inp_type_ids, inp_mask, subtokens_mask, start_word_ids, query_ids, is_first, is_last = batch
+        # We will store the output here
+        result = []
+
+        # Model's mode and device
+        mode = self.training
+        device = 'cuda' if torch.cuda.is_available() else 'cpu'
+        try:
+            # Switch model to evaluation mode
+            self.eval()
+            self = self.to(device)
+
+            infer_datalayer = self._setup_infer_dataloader(queries, batch_size, max_seq_length)
+
+            # store predictions for all queries in a single list
+            all_punct_preds = []
+            all_capit_preds = []
+
+            for batch in infer_datalayer:
+                input_ids, input_type_ids, input_mask, subtokens_mask = batch
+
                 punct_logits, capit_logits = self.forward(
-                    input_ids=inp_ids.to(d), token_type_ids=inp_type_ids.to(d), attention_mask=inp_mask.to(d),
+                    input_ids=input_ids.to(device),
+                    token_type_ids=input_type_ids.to(device),
+                    attention_mask=input_mask.to(device),
                 )
-                _res = self._transform_logit_to_prob_and_remove_margins_and_extract_word_probs(
-                    punct_logits, capit_logits, subtokens_mask, start_word_ids, margin, is_first, is_last, query_ids
-                )
-                punct_probs, capit_probs, start_word_ids = _res
-                for i, (q_i, start_word_id, bpp_i, bcp_i) in enumerate(
-                    zip(query_ids, start_word_ids, punct_probs, capit_probs)
-                ):
-                    for all_preds, acc_probs, b_probs_i in [
-                        (all_punct_preds, acc_punct_probs, bpp_i),
-                        (all_capit_preds, acc_capit_probs, bcp_i),
-                    ]:
-                        if acc_probs[q_i] is None:
-                            acc_probs[q_i] = b_probs_i
-                        else:
-                            all_preds[q_i], acc_probs[q_i] = self.move_acc_probs_to_token_preds(
-                                all_preds[q_i], acc_probs[q_i], start_word_id - len(all_preds[q_i]),
-                            )
-                            acc_probs[q_i] = self.update_accumulated_probabilities(acc_probs[q_i], b_probs_i)
-            for all_preds, acc_probs in [(all_punct_preds, acc_punct_probs), (all_capit_preds, acc_capit_probs)]:
-                for q_i, (pred, prob) in enumerate(zip(all_preds, acc_probs)):
-                    if prob is not None:
-                        all_preds[q_i], acc_probs[q_i] = self.move_acc_probs_to_token_preds(pred, prob, len(prob))
+
+                subtokens_mask = subtokens_mask > 0.5
+
+                punct_preds = [
+                    tensor2list(p_l[subtokens_mask[i]]) for i, p_l in enumerate(torch.argmax(punct_logits, axis=-1))
+                ]
+                capit_preds = [
+                    tensor2list(c_l[subtokens_mask[i]]) for i, c_l in enumerate(torch.argmax(capit_logits, axis=-1))
+                ]
+
+                all_punct_preds.extend(punct_preds)
+                all_capit_preds.extend(capit_preds)
+
+            punct_ids_to_labels = {v: k for k, v in self._cfg.punct_label_ids.items()}
+            capit_ids_to_labels = {v: k for k, v in self._cfg.capit_label_ids.items()}
+
+            queries = [q.strip().split() for q in queries]
             for i, query in enumerate(queries):
-                result.append(self.apply_punct_capit_predictions(query, all_punct_preds[i], all_capit_preds[i]))
+                punct_preds = all_punct_preds[i]
+                capit_preds = all_capit_preds[i]
+                if len(query) != len(punct_preds):
+                    logging.warning(
+                        f'Max sequence length of query {query} is set to {max_seq_length}. Truncating the input.'
+                    )
+
+                    # removing the end of phrase punctuation of the truncated segment
+                    punct_preds[-1] = 0
+                    max_len = len(punct_preds)
+                    query = query[:max_len]
+
+                query_with_punct_and_capit = ''
+                for j, word in enumerate(query):
+                    punct_label = punct_ids_to_labels[punct_preds[j]]
+                    capit_label = capit_ids_to_labels[capit_preds[j]]
+
+                    if capit_label != self._cfg.dataset.pad_label:
+                        word = word.capitalize()
+                    query_with_punct_and_capit += word
+                    if punct_label != self._cfg.dataset.pad_label:
+                        query_with_punct_and_capit += punct_label
+                    query_with_punct_and_capit += ' '
+
+                result.append(query_with_punct_and_capit.strip())
         finally:
             # set mode back to its original value
             self.train(mode=mode)
